requirements.txt
boto3==1.35.76
Flask==3.1.0
google_api_python_client==2.154.0
protobuf==5.29.1
PyYAML==6.0.2